pyproject.toml
[tool.poetry]
name = "camus-python"
version = "0.1.0"
description = ""
authors = ["Camus <jose.camus@globant.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
sqlalchemy = "^2.0.30"
psycopg2-binary = "^2.9.9"
pydantic = "^2.7.3"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
